School-Practicum_Web_programming
The project is the realization of a multi-page site on the topic of Hairdressing Studio Quatro ++ in HTML5 and CSS3. The web site contains realized gallery with the JQuery library lightbox, realization of a Google location with the geolocation, as well as the implementation of LESS CSS, and is a fully responsive site adapted to devices of different dimensions.
Technologies – HTML5, CSS3, Responsive design, JQuery, Geolocation